Unit tests moved from app module to addThermometerPresentation module
