+++ addThermometer/addThermometerPresentation/src/test/java/com/klewerro/mitemperature2alt/addThermometerPresentation/ConnectThermometerViewModelTest.kt
@@ -1,4 +1,4 @@
-package com.klewerro.mitemperature2alt.presentation.addThermometer
+package com.klewerro.mitemperature2alt.addThermometerPresentation
 
 import androidx.lifecycle.SavedStateHandle
 import app.cash.turbine.test
@@ -8,17 +8,14 @@ import assertk.assertions.isFalse
 import assertk.assertions.isNotNull
 import assertk.assertions.isNull
 import assertk.assertions.isTrue
-import com.klewerro.mitemperature2alt.addThermometerPresentation.ConnectThermometerEvent
-import com.klewerro.mitemperature2alt.addThermometerPresentation.ConnectThermometerViewModel
-import com.klewerro.mitemperature2alt.addThermometerPresentation.ConnectingStatus
 import com.klewerro.mitemperature2alt.coreTest.fake.FakePersistenceRepository
 import com.klewerro.mitemperature2alt.coreTest.fake.FakeThermometerRepository
 import com.klewerro.mitemperature2alt.coreTest.util.MainCoroutineExtension
 import com.klewerro.mitemperature2alt.coreTest.util.TestDispatchers
+import com.klewerro.mitemperature2alt.coreUi.UiConstants
 import com.klewerro.mitemperature2alt.domain.usecase.thermometer.connect.ConnectToDeviceUseCase
 import com.klewerro.mitemperature2alt.domain.usecase.thermometer.operations.ReadCurrentThermometerStatusUseCase
 import com.klewerro.mitemperature2alt.domain.usecase.thermometer.persistence.SaveThermometerUseCase
-import com.klewerro.mitemperature2alt.presentation.navigation.Route
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -50,7 +47,7 @@ class ConnectThermometerViewModelTest {
 
         savedStateHandle = SavedStateHandle(
             initialState = mapOf(
-                Route.ConnectDeviceRoutes.PARAM_ADDRESS to deviceAddress
+                UiConstants.NAV_PARAM_ADDRESS to deviceAddress
             )
         )
         connectToDeviceUseCase = ConnectToDeviceUseCase(fakeThermometerRepository)
